API

API

RipQP

RipQP

RipQP.ripqp — Function

ripqp(QM0; mode=:mono, regul=:classic, scaling=true, K=0,
      max_iter=200, ϵ_pdd=1e-8, ϵ_rb=1e-6, ϵ_rc=1e-6,
      max_iter32=40, ϵ_pdd32=1e-2, ϵ_rb32=1e-4, ϵ_rc32=1e-4,
      max_iter64=180, ϵ_pdd64=1e-4, ϵ_rb64=1e-5, ϵ_rc64=1e-5,
      ϵ_Δx=1e-16, ϵ_μ=1e-9, max_time=1200., display=true)

Minimize a convex quadratic problem. Algorithm stops when the criteria in pdd, rb, and rc are valid. Returns a GenericExecutionStats containing information about the solved problem.

QM0::QuadraticModel{T0}: problem to solve

mode::Symbol: should be :mono to use the mono-precision mode, or :multi to use the multi-precision mode (start in single precision and gradually transitions to T0)

regul::Symbol: if :classic, then the regularization is performed prior the factorization, if :dynamic, then the regularization is performed during the factorization, and if :none, no regularization is used

scaling::Bool: activate/deactivate scaling of A and Q in QM0

K::Int: number of centrality corrections (set to -1 for automatic computation)

max_iter::Int: maximum number of iterations

ϵ_pdd: primal-dual difference tolerance

ϵ_rb: primal tolerance

ϵ_rc: dual tolerance

max_iter32, ϵ_pdd32, ϵ_rb32, ϵ_rc32: same as max_iter, ϵ_pdd, ϵ_rb and ϵ_rc, but used for switching from single precision to double precision. They are only usefull when mode=:multi

max_iter64, ϵ_pdd64, ϵ_rb64, ϵ_rc64: same as max_iter, ϵ_pdd, ϵ_rb and ϵ_rc, but used for switching from double precision to quadruple precision. They are only usefull when mode=:multi and T0=Float128

ϵ_Δx: step tolerance for the current point estimate (note: this criterion is currently disabled)

ϵ_μ: duality measure tolerance (note: this criterion is currently disabled)

max_time: maximum time to solve QM0, in seconds

display::Bool: activate/deactivate iteration data display