RipQP

RipQP.ripqp — Function

stats = ripqp(QM :: QuadraticModel{T0}; iconf :: InputConfig{Int} = InputConfig(),
              itol :: InputTol{T0, Int} = InputTol(T0),
              display :: Bool = true) where {T0<:Real}

Minimize a convex quadratic problem. Algorithm stops when the criteria in pdd, rb, and rc are valid. Returns a GenericExecutionStats containing information about the solved problem.

QM :: QuadraticModel: problem to solve

iconf :: InputConfig{Int}: input RipQP configuration. See RipQP.InputConfig.

itol :: InputTol{T, Int} input Tolerances for the stopping criteria. See RipQP.InputTol.

display::Bool: activate/deactivate iteration data display

You can also use ripqp to solve a LLSModel:

stats = ripqp(LLS :: LLSModel{T0}; iconf :: InputConfig{Int} = InputConfig(),
              itol :: InputTol{T0, Int} = InputTol(T0),
              display :: Bool = true) where {T0<:Real}

Input Types

RipQP.InputConfig — Type

Type to specify the configuration used by RipQP.

mode :: Symbol: should be :mono to use the mono-precision mode, or :multi to use the multi-precision mode (start in single precision and gradually transitions to T0)

scaling :: Bool: activate/deactivate scaling of A and Q in QM0

presolve :: Bool : activate/deactivate presolve

normalize_rtol :: Bool = true : if true, the primal and dual tolerance for the stopping criteria are normalized by the initial primal and dual residuals

kc :: Int: number of centrality corrections (set to -1 for automatic computation)

refinement :: Symbol : should be :zoom to use the zoom procedure, :multizoom to use the zoom procedure with multi-precision (then mode should be :multi), ref to use the QP refinement procedure, multiref to use the QP refinement procedure with multi_precision (then mode should be :multi), or none to avoid refinements

sp :: SolverParams : choose a solver to solve linear systems that occurs at each iteration and during the initialization, see RipQP.SolverParams

solve_method :: Symbol : method used to solve the system at each iteration, use solve_method = :PC to use the Predictor-Corrector algorithm (default), and use solve_method = :IPF to use the Infeasible Path Following algorithm

history :: Bool : set to true to return the primal and dual norm histories, the primal-dual relative difference history, and the number of products if using a Krylov method in the solver_specific field of the GenericExecutionStats

w :: SystemWrite: configure writing of the systems to solve (no writing is done by default), see RipQP.SystemWrite

The constructor

iconf = InputConfig(; mode :: Symbol = :mono, scaling :: Bool = true, 
                    normalize_rtol :: Bool = true, kc :: I = 0, 
                    refinement :: Symbol = :none, max_ref :: I = 0, 
                    sp :: SolverParams = K2LDLParams(),
                    solve_method :: Symbol = :PC,
                    history :: Bool = false, 
                    w :: SystemWrite = SystemWrite()) where {I<:Integer}

returns a InputConfig struct that shall be used to solve the input QuadraticModel with RipQP.

RipQP.InputTol — Type

Type to specify the tolerances used by RipQP.

max_iter :: Int: maximum number of iterations

ϵ_pdd: relative primal-dual difference tolerance

ϵ_rb: primal tolerance

ϵ_rc: dual tolerance

max_iter32, ϵ_pdd32, ϵ_rb32, ϵ_rc32: same as max_iter, ϵ_pdd, ϵ_rb and ϵ_rc, but used for switching from single precision to double precision. They are only usefull when mode=:multi

max_iter64, ϵ_pdd64, ϵ_rb64, ϵ_rc64: same as max_iter, ϵ_pdd, ϵ_rb and ϵ_rc, but used for switching from double precision to quadruple precision. They are only usefull when mode=:multi and T0=Float128

ϵ_rbz : primal transition tolerance for the zoom procedure, (used only if refinement=:zoom)

ϵ_Δx: step tolerance for the current point estimate (note: this criterion is currently disabled)

ϵ_μ: duality measure tolerance (note: this criterion is currently disabled)

max_time: maximum time to solve the QP

The constructor

itol = InputTol(::Type{T};
                max_iter :: I = 200, max_iter32 :: I = 40, max_iter64 :: I = 180, 
                ϵ_pdd :: T = 1e-8, ϵ_pdd32 :: T = 1e-2, ϵ_pdd64 :: T = 1e-4, 
                ϵ_rb :: T = 1e-6, ϵ_rb32 :: T = 1e-4, ϵ_rb64 :: T = 1e-5, ϵ_rbz :: T = 1e-3,
                ϵ_rc :: T = 1e-6, ϵ_rc32 :: T = 1e-4, ϵ_rc64 :: T = 1e-5,
                ϵ_Δx :: T = 1e-16, ϵ_μ :: T = 1e-9) where {T<:Real, I<:Integer}

InputTol(; kwargs...) = InputTol(Float64; kwargs...)

returns a InputTol struct that initializes the stopping criteria for RipQP. The 32 and 64 characters refer to the stopping criteria in :multi mode for the transitions from Float32 to Float64 and Float64 to Float128 (if the input QuadraticModel is in Float128) respectively.

RipQP.SystemWrite — Type

Type to write the matrix (.mtx format) and the right hand side (.rhs format) of the system to solve at each iteration.

write::Bool: activate/deactivate writing of the system

name::String: name of the sytem to solve

kfirst::Int: first iteration where a system should be written

kgap::Int: iteration gap between two problem writings

The constructor

SystemWrite(; write::Bool = false, name::String = "", 
            kfirst::Int = 0, kgap::Int = 1)

returns a SystemWrite structure that should be used to tell RipQP to save the system. See the tutorial for more information.

Solvers

RipQP.SolverParams — Type

Abstract type for tuning the parameters of the different solvers. Each solver has its own SolverParams type.

The SolverParams currently implemented within RipQP are:

RipQP.K2LDLParams

RipQP.K2_5LDLParams

RipQP.K2LDLParams — Type

Type to use the K2 formulation with a LDLᵀ factorization, using the package LDLFactorizations.jl. The outer constructor

sp = K2LDLParams(; regul = :classic, ρ0 = sqrt(eps()) * 1e5, δ0 = sqrt(eps()) * 1e5)

creates a RipQP.SolverParams that should be used to create a RipQP.InputConfig. regul = :dynamic uses a dynamic regularization (the regularization is only added if the LDLᵀ factorization encounters a pivot that has a small magnitude). regul = :none uses no regularization (not recommended). When regul = :classic, the parameters ρ0 and δ0 are used to choose the initial regularization values.

RipQP.K2_5LDLParams — Type

Type to use the K2.5 formulation with a LDLᵀ factorization, using the package LDLFactorizations.jl. The outer constructor

sp = K2_5LDLParams(; regul = :classic, ρ0 = sqrt(eps()) * 1e5, δ0 = sqrt(eps()) * 1e5)

creates a RipQP.SolverParams that should be used to create a RipQP.InputConfig. regul = :dynamic uses a dynamic regularization (the regularization is only added if the LDLᵀ factorization encounters a pivot that has a small magnitude). regul = :none uses no regularization (not recommended). When regul = :classic, the parameters ρ0 and δ0 are used to choose the initial regularization values.

RipQP.K2KrylovParams — Type

Type to use the K2 formulation with a Krylov method, using the package Krylov.jl. The outer constructor

K2KrylovParams(; uplo = :L, kmethod = :minres, preconditioner = :Identity,
               rhs_scale = true, form_mat = false,
               atol0 = 1.0e-4, rtol0 = 1.0e-4, 
               atol_min = 1.0e-10, rtol_min = 1.0e-10,
               ρ0 = sqrt(eps()) * 1e5, δ0 = sqrt(eps()) * 1e5, 
               ρ_min = 1e2 * sqrt(eps()), δ_min = 1e2 * sqrt(eps()),
               memory = 20)

creates a RipQP.SolverParams that should be used to create a RipQP.InputConfig. The available methods are:

:minres

:minres_qlp

:symmlq

The list of available preconditioners for this solver is displayed here: RipQP.PreconditionerData.

RipQP.K2_5KrylovParams — Type

Type to use the K2.5 formulation with a Krylov method, using the package Krylov.jl. The outer constructor

K2_5KrylovParams(; uplo = :L, kmethod = :minres, preconditioner = :Identity,
                 rhs_scale = true,
                 atol0 = 1.0e-4, rtol0 = 1.0e-4, 
                 atol_min = 1.0e-10, rtol_min = 1.0e-10,
                 ρ0 = sqrt(eps()) * 1e5, δ0 = sqrt(eps()) * 1e5, 
                 ρ_min = 1e2 * sqrt(eps()), δ_min = 1e2 * sqrt(eps()),
                 mem = 20)

creates a RipQP.SolverParams that should be used to create a RipQP.InputConfig. The available methods are:

:minres

:minres_qlp

:symmlq

The list of available preconditioners for this solver is displayed here: RipQP.PreconditionerData

RipQP.K2StructuredParams — Type

Type to use the K2 formulation with a structured Krylov method, using the package Krylov.jl. This only works for solving Linear Problems. The outer constructor

K2StructuredParams(; uplo = :L, kmethod = :trimr, rhs_scale = true, 
                   atol0 = 1.0e-4, rtol0 = 1.0e-4,
                   atol_min = 1.0e-10, rtol_min = 1.0e-10, 
                   ρ_min = 1e2 * sqrt(eps()), δ_min = 1e2 * sqrt(eps()),
                   mem = 20)

creates a RipQP.SolverParams that should be used to create a RipQP.InputConfig. The available methods are:

:tricg

:trimr

:gpmr

The list of available preconditioners for this solver is displayed here: RipQP.PreconditionerData. The mem argument sould be used only with gpmr.

RipQP.K2_5StructuredParams — Type

Type to use the K2.5 formulation with a structured Krylov method, using the package Krylov.jl. This only works for solving Linear Problems. The outer constructor

K2_5StructuredParams(; uplo = :L, kmethod = :trimr, rhs_scale = true,
                     atol0 = 1.0e-4, rtol0 = 1.0e-4,
                     atol_min = 1.0e-10, rtol_min = 1.0e-10,
                     ρ0 = sqrt(eps()) * 1e5, δ0 = sqrt(eps()) * 1e5,
                     ρ_min = 1e2 * sqrt(eps()), δ_min = 1e2 * sqrt(eps()))

creates a RipQP.SolverParams that should be used to create a RipQP.InputConfig. The available methods are:

:tricg

:trimr

:gpmr

The list of available preconditioners for this solver is displayed here: RipQP.PreconditionerData. The mem argument sould be used only with gpmr.

RipQP.K3KrylovParams — Type

Type to use the K3 formulation with a Krylov method, using the package Krylov.jl. The outer constructor

K3KrylovParams(; uplo = :L, kmethod = :qmr, preconditioner = :Identity,
               rhs_scale = true,
               atol0 = 1.0e-4, rtol0 = 1.0e-4,
               atol_min = 1.0e-10, rtol_min = 1.0e-10,
               ρ0 = sqrt(eps()) * 1e5, δ0 = sqrt(eps()) * 1e5,
               ρ_min = 1e3 * sqrt(eps()), δ_min = 1e4 * sqrt(eps()),
               mem = 20)

creates a RipQP.SolverParams that should be used to create a RipQP.InputConfig. The available methods are:

:qmr

:bicgstab

:usymqr

RipQP.K3SKrylovParams — Type

Type to use the K3S formulation with a Krylov method, using the package Krylov.jl. The outer constructor

K3SKrylovParams(; uplo = :L, kmethod = :minres, preconditioner = :Identity,
                 rhs_scale = true,
                 atol0 = 1.0e-4, rtol0 = 1.0e-4,
                 atol_min = 1.0e-10, rtol_min = 1.0e-10,
                 ρ0 = sqrt(eps()) * 1e5, δ0 = sqrt(eps()) * 1e5,
                 ρ_min = 1e3 * sqrt(eps()), δ_min = 1e4 * sqrt(eps()),
                 mem = 20)

creates a RipQP.SolverParams that should be used to create a RipQP.InputConfig. The available methods are:

:minres

:minres_qlp

:symmlq

RipQP.K3SStructuredParams — Type

Type to use the K3S formulation with a Krylov method, using the package Krylov.jl. The outer constructor

K3SStructuredParams(; uplo = :U, kmethod = :trimr, rhs_scale = true,
                     atol0 = 1.0e-4, rtol0 = 1.0e-4, 
                     atol_min = 1.0e-10, rtol_min = 1.0e-10,
                     ρ0 =  sqrt(eps()) * 1e3, δ0 = sqrt(eps()) * 1e4,
                     ρ_min = 1e4 * sqrt(eps()), δ_min = 1e4 * sqrt(eps()),
                     mem = 20)

creates a RipQP.SolverParams that should be used to create a RipQP.InputConfig. The available methods are:

:tricg

:trimr

:gpmr

The mem argument sould be used only with gpmr.

RipQP.K3_5KrylovParams — Type

Type to use the K3.5 formulation with a Krylov method, using the package Krylov.jl. The outer constructor

K3_5KrylovParams(; uplo = :L, kmethod = :minres, preconditioner = :Identity,
                 rhs_scale = true,
                 atol0 = 1.0e-4, rtol0 = 1.0e-4,
                 atol_min = 1.0e-10, rtol_min = 1.0e-10,
                 ρ0 = sqrt(eps()) * 1e5, δ0 = sqrt(eps()) * 1e5,
                 ρ_min = 1e3 * sqrt(eps()), δ_min = 1e4 * sqrt(eps()),
                 mem = 20)

creates a RipQP.SolverParams that should be used to create a RipQP.InputConfig. The available methods are:

:minres

:minres_qlp

:symmlq

RipQP.K3_5StructuredParams — Type

Type to use the K3.5 formulation with a Krylov method, using the package Krylov.jl. The outer constructor

K3_5StructuredParams(; uplo = :U, kmethod = :trimr, rhs_scale = true,
                     atol0 = 1.0e-4, rtol0 = 1.0e-4, 
                     atol_min = 1.0e-10, rtol_min = 1.0e-10,
                     ρ0 =  sqrt(eps()) * 1e3, δ0 = sqrt(eps()) * 1e4,
                     ρ_min = 1e4 * sqrt(eps()), δ_min = 1e4 * sqrt(eps()),
                     mem = 20)

creates a RipQP.SolverParams that should be used to create a RipQP.InputConfig. The available methods are:

:tricg

:trimr

:gpmr

The mem argument sould be used only with gpmr.

RipQP.K1KrylovParams — Type

Type to use the K1 formulation with a Krylov method, using the package Krylov.jl. The outer constructor

K1KrylovParams(; uplo = :L, kmethod = :cg, preconditioner = :Identity,
               rhs_scale = true,
               atol0 = 1.0e-4, rtol0 = 1.0e-4, 
               atol_min = 1.0e-10, rtol_min = 1.0e-10,
               ρ0 = sqrt(eps()) * 1e5, δ0 = sqrt(eps()) * 1e5, 
               ρ_min = 1e2 * sqrt(eps()), δ_min = 1e2 * sqrt(eps()),
               mem = 20)

creates a RipQP.SolverParams that should be used to create a RipQP.InputConfig. The available methods are:

:cg

:cg_lanczos

:cr

:minres

:minres_qlp

:symmlq

RipQP.K1CholDenseParams — Type

Type to use the K1 formulation with a dense Cholesky factorization. The input QuadraticModel should have lcon .== ucon. The outer constructor

sp = K1CholDenseParams(; ρ0 = sqrt(eps()) * 1e5, δ0 = sqrt(eps()) * 1e5)

creates a RipQP.SolverParams that should be used to create a RipQP.InputConfig.

RipQP.K2LDLDenseParams — Type

Type to use the K2 formulation with a LDLᵀ factorization. The outer constructor

sp = K2LDLDenseParams(; ρ0 = sqrt(eps()) * 1e5, δ0 = sqrt(eps()) * 1e5)

creates a RipQP.SolverParams that should be used to create a RipQP.InputConfig.

RipQP.K1_1StructuredParams — Type

Type to use the K1.1 formulation with a structured Krylov method, using the package Krylov.jl. This only works for solving Linear Problems. The outer constructor

K1_1StructuredParams(; uplo = :L, kmethod = :lsqr, rhs_scale = true,
                     atol0 = 1.0e-4, rtol0 = 1.0e-4,
                     atol_min = 1.0e-10, rtol_min = 1.0e-10, 
                     ρ_min = 1e3 * sqrt(eps()), δ_min = 1e4 * sqrt(eps()),
                     mem = 20)

creates a RipQP.SolverParams that should be used to create a RipQP.InputConfig. The available methods are:

:lslq

:lsqr

:lsmr

RipQP.K1_2StructuredParams — Type

Type to use the K1.2 formulation with a structured Krylov method, using the package Krylov.jl. This only works for solving Linear Problems. The outer constructor

K1_2StructuredParams(; uplo = :L, kmethod = :craig, rhs_scale = true,
                     atol0 = 1.0e-4, rtol0 = 1.0e-4,
                     atol_min = 1.0e-10, rtol_min = 1.0e-10, 
                     ρ_min = 1e3 * sqrt(eps()), δ_min = 1e4 * sqrt(eps()),
                     mem = 20)

creates a RipQP.SolverParams that should be used to create a RipQP.InputConfig. The available methods are:

:lnlq

:craig

:craigmr

Preconditioners

RipQP.PreconditionerData — Type

Abstract type that defines preconditioners. The available preconditioners are:

Identity

Jacobi

Equilibration