API

API

RipQP

Input Types

RipQP

RipQP.ripqp — Function

stats = ripqp(QM :: QuadraticModel; iconf :: InputConfig{Int} = InputConfig(), 
              itol :: InputTol{Tu, Int} = InputTol(), 
              display :: Bool = true) where {Tu<:Real}

Minimize a convex quadratic problem. Algorithm stops when the criteria in pdd, rb, and rc are valid. Returns a GenericExecutionStats containing information about the solved problem.

QM :: QuadraticModel: problem to solve

iconf :: InputConfig{Int}: input RipQP configuration. See InputConfig{I}.

itol :: InputTol{T, Int} input Tolerances for the stopping criteria. See InputTol{T, I}.

display::Bool: activate/deactivate iteration data display

Input Types

RipQP.InputConfig — Type

Type to specify the configuration used by RipQP.

mode :: Symbol: should be :mono to use the mono-precision mode, or :multi to use the multi-precision mode (start in single precision and gradually transitions to T0)

regul :: Symbol: if :classic, then the regularization is performed prior the factorization, if :dynamic, then the regularization is performed during the factorization, and if :none, no regularization is used

scaling :: Bool: activate/deactivate scaling of A and Q in QM0

normalize_rtol :: Bool = true : if true, the primal and dual tolerance for the stopping criteria are normalized by the initial primal and dual residuals

kc :: Int: number of centrality corrections (set to -1 for automatic computation)

refinement :: Symbol : should be :zoom to use the zoom procedure, :multizoom to use the zoom procedure with multi-precision (then mode should be :multi), ref to use the QP refinement procedure, multiref to use the QP refinement procedure with multi_precision (then mode should be :multi), or none to avoid refinements

solver :: Symbol : choose a solver to solve linear systems that occurs at each iteration and during the initialization (the Symbol should correspond to the specific PreallocatedData used by the solver)

solve_method :: Symbol : used to solve the system at each iteration

The constructor

iconf = InputConfig(; mode :: Symbol = :mono, regul :: Symbol = :classic, 
                    scaling :: Bool = true, normalize_rtol :: Bool = true, 
                    kc :: I = 0, refinement :: Symbol = :none, max_ref :: I = 0, 
                    solver :: Symbol = :K2, solve_method :: Symbol = :PC) where {I<:Integer}

returns a InputConfig struct that shall be used to solve the input QuadraticModel with RipQP.

RipQP.InputTol — Type

Type to specify the tolerances used by RipQP.

max_iter :: Int: maximum number of iterations

ϵ_pdd: relative primal-dual difference tolerance

ϵ_rb: primal tolerance

ϵ_rc: dual tolerance

max_iter32, ϵ_pdd32, ϵ_rb32, ϵ_rc32: same as max_iter, ϵ_pdd, ϵ_rb and ϵ_rc, but used for switching from single precision to double precision. They are only usefull when mode=:multi

max_iter64, ϵ_pdd64, ϵ_rb64, ϵ_rc64: same as max_iter, ϵ_pdd, ϵ_rb and ϵ_rc, but used for switching from double precision to quadruple precision. They are only usefull when mode=:multi and T0=Float128

ϵ_rbz : primal transition tolerance for the zoom procedure, (used only if refinement=:zoom)

ϵ_Δx: step tolerance for the current point estimate (note: this criterion is currently disabled)

ϵ_μ: duality measure tolerance (note: this criterion is currently disabled)

max_time: maximum time to solve the QP

The constructor

itol = InputTol(;max_iter :: I = 200, max_iter32 :: I = 40, max_iter64 :: I = 180, 
                 ϵ_pdd :: T = 1e-8, ϵ_pdd32 :: T = 1e-2, ϵ_pdd64 :: T = 1e-4, 
                 ϵ_rb :: T = 1e-6, ϵ_rb32 :: T = 1e-4, ϵ_rb64 :: T = 1e-5, ϵ_rbz :: T = 1e-3,
                 ϵ_rc :: T = 1e-6, ϵ_rc32 :: T = 1e-4, ϵ_rc64 :: T = 1e-5,
                 ϵ_Δx :: T = 1e-16, ϵ_μ :: T = 1e-9) where {T<:Real, I<:Integer}

returns a InputTol struct that initializes the stopping criteria for RipQP. The 32 and 64 characters refer to the stopping criteria in :multi mode for the transitions from Float32 to Float64 and Float64 to Float128 (if the input QuadraticModel is in Float128) respectively.